Project restructure: pages folder and HomeComponent inside of it.
