+++ src/app/app.component.html
@@ -1,9 +1,3 @@
 <web-header></web-header>
-
-<web-hero></web-hero>
-<web-experience></web-experience>
-<web-projects></web-projects>
-<web-about-me></web-about-me>
-<web-contact></web-contact>
-
+<web-home></web-home>
 <web-footer></web-footer>
