+++ src/app/app.component.ts
@@ -1,11 +1,7 @@
 import { Component } from '@angular/core';
 import { HeaderComponent } from './layout/header/header.component';
 import { FooterComponent } from "./layout/footer/footer.component";
-import { HeroComponent } from "./layout/hero/hero.component";
-import { ExperienceComponent } from "./layout/experience/experience.component";
-import { ProjectsComponent } from './layout/projects/projects.component';
-import { AboutMeComponent } from './layout/about-me/about-me.component';
-import { ContactComponent } from './layout/contact/contact.component';
+import { HomeComponent } from './pages/home/home.component';
 
 @Component({
   selector: 'app-root',
@@ -13,12 +9,8 @@ import { ContactComponent } from './layout/contact/contact.component';
   imports: [
     HeaderComponent,
     FooterComponent,
-    HeroComponent,
-    ExperienceComponent,
-    ProjectsComponent,
-    AboutMeComponent,
-    ContactComponent
-],
+    HomeComponent
+  ],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
 })
