+++ src/app/pages/home/contact/contact.component.ts
@@ -1,5 +1,5 @@
 import { Component, inject, OnInit } from '@angular/core';
-import { ResourceService } from '../../resource.service';
+import { ResourceService } from '../../../resource.service';
 
 @Component({
   selector: 'web-contact',
