+++ src/app/pages/home/home.component.html
@@ -1,0 +1,5 @@
+<web-hero></web-hero>
+<web-experience></web-experience>
+<web-projects></web-projects>
+<web-about-me></web-about-me>
+<web-contact></web-contact>
