+++ src/app/pages/home/projects/projects.component.ts
